requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
yt-dlp==2024.7.25
